--- iris-mpc-cpu/src/indexation/genesis/actors1/mod.rs
@@ -1,9 +1,0 @@
-mod batch_generator;
-mod graph_indexer;
-mod shares_fetcher;
-mod supervisor;
-
-pub use batch_generator::BatchGenerator;
-pub use graph_indexer::GraphIndexer;
-pub use shares_fetcher::SharesFetcher;
-pub use supervisor::Supervisor;
