+++ iris-mpc-cpu/src/indexation/genesis/mod.rs
@@ -1,12 +1,10 @@
 mod actors;
-mod actors1;
 mod errors;
 mod messages;
 mod types;
 mod utils;
 
 pub use actors::Supervisor;
-pub use actors1::Supervisor as Supervisor1;
 pub use errors::IndexationError;
 pub use messages::OnBeginIndexation;
 pub use messages::OnEndIndexation;
